requirements.txt
django
django-allauth
django-channels
channels-redis
django-crispy-forms
pillow
# packages not directly related to Django
xhtml2pdf
